+++ cpp/oneapi/dal/algo/pca/backend/cpu/finalize_train_kernel_cov.cpp
@@ -18,7 +18,6 @@
 #include <daal/src/algorithms/covariance/covariance_hyperparameter_impl.h>
 #include "daal/src/algorithms/covariance/covariance_kernel.h"
 
-#include "oneapi/dal/algo/pca/backend/common.hpp"
 #include "oneapi/dal/algo/pca/backend/cpu/finalize_train_kernel.hpp"
 #include "oneapi/dal/backend/interop/common.hpp"
