+++ cpp/oneapi/dal/algo/pca/backend/cpu/train_kernel.hpp
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include "oneapi/dal/algo/pca/train_types.hpp"
 #include "oneapi/dal/backend/dispatcher.hpp"
 
 namespace oneapi::dal::pca::backend {
