+++ cpp/oneapi/dal/algo/pca/backend/cpu/train_kernel_precomputed.cpp
@@ -16,7 +16,6 @@
 
 #include <daal/src/algorithms/pca/pca_dense_correlation_batch_kernel.h>
 
-#include "oneapi/dal/algo/pca/backend/common.hpp"
 #include "oneapi/dal/algo/pca/backend/cpu/train_kernel.hpp"
 #include "oneapi/dal/backend/interop/common.hpp"
 #include "oneapi/dal/backend/interop/error_converter.hpp"
