+++ cpp/oneapi/dal/algo/pca/backend/gpu/finalize_train_kernel_svd_dpc.cpp
@@ -15,12 +15,12 @@
 *******************************************************************************/
 
 #include "oneapi/dal/algo/pca/backend/gpu/finalize_train_kernel.hpp"
-#include "oneapi/dal/algo/pca/backend/common.hpp"
+
 #include "oneapi/dal/backend/primitives/lapack.hpp"
 #include "oneapi/dal/backend/primitives/reduction.hpp"
 #include "oneapi/dal/backend/primitives/stat.hpp"
 #include "oneapi/dal/backend/primitives/utils.hpp"
-#include "oneapi/dal/algo/pca/backend/sign_flip.hpp"
+#include "oneapi/dal/algo/pca/sign_flip.hpp"
 #include "oneapi/dal/table/row_accessor.hpp"
 namespace oneapi::dal::pca::backend {
