+++ cpp/oneapi/dal/algo/pca/backend/gpu/train_kernel_cov_dpc.cpp
@@ -16,8 +16,8 @@
 
 #include "oneapi/dal/algo/pca/backend/gpu/train_kernel.hpp"
 #include "oneapi/dal/algo/pca/backend/gpu/train_kernel_cov_impl.hpp"
-#include "oneapi/dal/algo/pca/backend/common.hpp"
-#include "oneapi/dal/algo/pca/backend/sign_flip.hpp"
+
+#include "oneapi/dal/algo/pca/sign_flip.hpp"
 #include "oneapi/dal/table/row_accessor.hpp"
 #include "oneapi/dal/detail/policy.hpp"
 #include "oneapi/dal/detail/common.hpp"
