+++ cpp/oneapi/dal/algo/pca/backend/gpu/train_kernel_cov_impl_dpc.cpp
@@ -23,8 +23,8 @@
 #include "oneapi/dal/backend/primitives/reduction.hpp"
 #include "oneapi/dal/backend/primitives/stat.hpp"
 #include "oneapi/dal/backend/primitives/blas.hpp"
-#include "oneapi/dal/algo/pca/backend/common.hpp"
-#include "oneapi/dal/algo/pca/backend/sign_flip.hpp"
+
+#include "oneapi/dal/algo/pca/sign_flip.hpp"
 
 #ifdef ONEDAL_DATA_PARALLEL
